Adj demo

+++ src/com/axiastudio/pypapi/demo/entities/Book.java
@@ -34,7 +34,7 @@ public class Book implements Serializable {
     @GeneratedValue(strategy=GenerationType.TABLE)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "title")
     private String title;
     @Column(name = "description")
@@ -46,15 +46,15 @@ public class Book implements Serializable {
     public Book() {
     }
 
-    public Book(Integer id) {
+    public Book(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
