+++ src/com/axiastudio/pypapi/demo/entities/Loan.java
@@ -21,7 +21,7 @@ public class Loan implements Serializable {
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "fromdate")
     @Temporal(TemporalType.DATE)
     private Date fromdate;
@@ -40,15 +40,15 @@ public class Loan implements Serializable {
     public Loan() {
     }
 
-    public Loan(Integer id) {
+    public Loan(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
